replace vp
